Modification of filename_blacklist behavior.
filename_blacklist should be able to accept a simple string
specifying an exact filename you would like to ignore.

example:
"filename_blacklist": [".DS_Store", "Thumbs.db"],

I believe this is a more intuitive use of a filename blacklist and the
ability to add a dict with advanced functionality still exists.