Add leave_symlink option
Each time a file is moved/renamed a symlink is left behind, named after the
just-moved file, pointing to the new location. Since at present the file is
moved twice (once for the rename, again for the move) two symlinks are made.
This isn't ideal in my opinion but it'll do for now.